requirements.txt
mcp-server
python-dotenv
mem0
supabase-py 